+++ packages/flutter_adaptive_cards_fs/lib/src/cards/elements/compound_button.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_adaptive_cards_fs/src/adaptive_mixins.dart';
 import 'package:flutter_adaptive_cards_fs/src/additional.dart';
-import 'package:flutter_adaptive_cards_fs/src/riverpod_providers.dart';
 import 'package:flutter_adaptive_cards_fs/src/utils/adaptive_image_utils.dart';
 import 'package:flutter_adaptive_cards_fs/src/utils/utils.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class AdaptiveCompoundButton extends StatefulWidget
     with AdaptiveElementWidgetMixin {
@@ -40,9 +38,7 @@ class AdaptiveCompoundButtonState extends State<AdaptiveCompoundButton>
 
   @override
   Widget build(BuildContext context) {
-    final resolver = ProviderScope.containerOf(
-      context,
-    ).read(styleReferenceResolverProvider);
+    final resolver = styleResolver;
     return Visibility(
       visible: isVisible,
       child: SeparatorElement(