+++ packages/flutter_adaptive_cards_fs/lib/src/cards/inputs/date.dart
@@ -3,9 +3,7 @@ import 'dart:developer' as developer;
 import 'package:flutter/material.dart';
 import 'package:flutter_adaptive_cards_fs/src/adaptive_mixins.dart';
 import 'package:flutter_adaptive_cards_fs/src/additional.dart';
-import 'package:flutter_adaptive_cards_fs/src/riverpod_providers.dart';
 import 'package:flutter_adaptive_cards_fs/src/utils/utils.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:intl/intl.dart';
 
 ///
@@ -122,9 +120,7 @@ class AdaptiveDateInputState extends State<AdaptiveDateInput>
                     borderSide: BorderSide(width: 1),
                   ),
                   filled: true,
-                  fillColor: ProviderScope.containerOf(context)
-                      .read(styleReferenceResolverProvider)
-                      .resolveInputBackgroundColor(
+                  fillColor: styleResolver.resolveInputBackgroundColor(
                         context: context,
                         style: null,
                       ),